+++ refena/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.36.0
 
 - feat: add `RefreshAction`, a new class of action that reduces the boilerplate when refreshing `ReduxNotifier`s with data type `AsyncValue`
+- **BREAKING**: remove `ref.watchWithPrev`, `AsyncValue` already contains the previous value
 
 ## 0.35.0